- DRY tests - Fix typo on 'CELL_CSS_SELCTOR' (~> CELL_CSS_SELECTOR) - Remove useless HTMLDivElement from ResizerSupervisor.tsx - Add useCallback into setFunctionName function (InvocationExpression.tsx) - Bring noHandlerMenu back (ContextExpression.tsx)
